Add instrumentation tests
- Adds tests that run against an actual repo (set via the secrets in
this repo) using the latest version of the action. These tests use a
script to programmatically delete all releases and tags in the repo and
then create releases and tags based on the arguments passed into it.
Then, it uses the Action in order to test that it accurately deletes
those tags and releases based on different arguments.